imageMerge/requirements.txt
pillow
imageio
pygifsicle